feat: ensure cart product decrease quantity

+++ src/actions/decrease-cart-product-quantity/index.tsx
@@ -1,22 +1,26 @@
-'use server'
+"use server";
 
+import { eq } from "drizzle-orm";
 import { headers } from "next/headers";
 import { z } from "zod";
 
 import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
-import { eq } from "drizzle-orm";
 
-const removeProductFromCartSchema = z.object({
+const decreaseCartProductQuantitySchema = z.object({
   cartItemId: z.uuid(),
 });
 
-export type RemoveProductFromCartSchema = z.infer<typeof removeProductFromCartSchema>;
+type DecreaseCartProductQuantitySchema = z.infer<
+  typeof decreaseCartProductQuantitySchema
+>;
 
-export async function removeProductFromCart(data: RemoveProductFromCartSchema) {
+export const decreaseCartProductQuantity = async (
+  data: DecreaseCartProductQuantitySchema,
+) => {
   // valida os dados
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
 
   // valida se o usuário está autenticado
   const session = await auth.api.getSession({
@@ -32,20 +36,30 @@ export async function removeProductFromCart(data: RemoveProductFromCartSchema) {
     where: (cartItem, { eq }) => eq(cartItem.id, data.cartItemId),
     with: {
       cart: true,
-    }
+    },
   });
 
   // Verifica se existe carrinho
   if (!cartItem) {
-    throw new Error("Product not found in cart");
+    throw new Error("Cart item not found");
   }
 
   // Verifica se o id do produto pertence a um carrinho e pertence ao usuario
-  const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id
+  const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id;
 
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
-}
+  // Se a quantidade for 1 remove o produto
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  // Diminui a quantidade se tiver mais do que uma
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
+};
